#10 | create separate directory for stylesheets
