+++ screens/HomeScreen.js
@@ -26,6 +26,8 @@ import {
 
 import { MonoText } from '../components/StyledText';
 
+import styles from '../styles/homeScreen/style';
+
 export default class HomeScreen extends React.Component {
   static route = {
     navigationBar: {
@@ -85,7 +87,7 @@ export default class HomeScreen extends React.Component {
             />
           </Item>
           <Button
-style={{ flex: 0.1 }} rounded info
+            style={{ flex: 0.1 }} rounded info
             onPress={this._handleHelpPress}
           >
             <Text style={{ textAlign: 'center', paddingLeft: 5 }}>Send</Text>
@@ -143,104 +145,3 @@ style={{ flex: 0.1 }} rounded info
     });
   };
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 0.2,
-    backgroundColor: '#fff',
-    flexDirection: 'row',
-    marginTop: 5,
-    marginHorizontal: 5,
-  },
-  responseContainer: {
-    borderRadius: 4,
-    borderWidth: 0.5,
-    borderColor: '#d6d7da',
-    marginRight: 20,
-    marginLeft: 20,
-    padding: 10,
-  },
-  developmentModeText: {
-    marginBottom: 20,
-    color: 'rgba(0,0,0,0.4)',
-    fontSize: 15,
-    textAlign: 'center',
-  },
-  contentContainer: {
-    paddingTop: 80,
-  },
-  welcomeContainer: {
-    alignItems: 'center',
-    marginTop: 10,
-    marginBottom: 20,
-  },
-  welcomeImage: {
-    width: 140,
-    height: 38,
-    resizeMode: 'contain',
-    marginTop: 3,
-    marginLeft: -10,
-  },
-  getStartedContainer: {
-    alignItems: 'center',
-    marginRight: 20,
-    marginTop: 10,
-    marginLeft: 100,
-  },
-  homeScreenFilename: {
-    marginVertical: 7,
-  },
-  codeHighlightText: {
-    color: 'rgba(96,100,109, 0.8)',
-  },
-  codeHighlightContainer: {
-    backgroundColor: 'rgba(0,0,0,0.05)',
-    borderRadius: 3,
-    paddingHorizontal: 4,
-  },
-  getStartedText: {
-    fontSize: 17,
-    color: 'rgba(96,100,109, 1)',
-    lineHeight: 23,
-    textAlign: 'center',
-  },
-  tabBarInfoContainer: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    ...Platform.select({
-      ios: {
-        shadowColor: 'black',
-        shadowOffset: { height: -3 },
-        shadowOpacity: 0.1,
-        shadowRadius: 3,
-      },
-      android: {
-        elevation: 20,
-      },
-    }),
-    alignItems: 'center',
-    backgroundColor: '#fbfbfb',
-    paddingVertical: 20,
-  },
-  tabBarInfoText: {
-    fontSize: 17,
-    color: 'rgba(96,100,109, 1)',
-    textAlign: 'center',
-  },
-  navigationFilename: {
-    marginTop: 5,
-  },
-  helpContainer: {
-    marginTop: 15,
-    alignItems: 'center',
-  },
-  helpLink: {
-    paddingVertical: 15,
-  },
-  helpLinkText: {
-    fontSize: 14,
-    color: '#2e78b7',
-  },
-});
